competed coding

+++ migrations/sqls/20220124032725-orders-products-table-up.sql
@@ -2,6 +2,6 @@
 CREATE TABLE orders_products(
     id SERIAL PRIMARY KEY,
     quantity INTEGER,
-    order_id BIGINT REFERENCES orders(id),
-    product_id BIGINT REFERENCES products(id)
+    order_id INTEGER REFERENCES orders(id),
+    product_id INTEGER REFERENCES products(id)
 );
